// apps/derrops-portal/src/client/derrops-cloud/models/api-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Derrops API
 * Internal API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OaSpecRef } from './oa-spec-ref';
// May contain unused imports in some cases
// @ts-ignore
import type { VersionFetchState } from './version-fetch-state';

export interface ApiEntity {
    'id': string;
    'tenantId': string;
    'name': string;
    'description'?: string | null;
    /**
     * External URL where the spec can be found or fetched from
     */
    'externalUrl'?: string | null;
    /**
     * Spec format type
     */
    'specType': string;
    /**
     * private  — tenant owns and manages this spec. platform — tenant has adopted a Derrops-managed spec from the global catalogue.
     */
    'managementMode': ApiEntityManagementModeEnum;
    'createdAt': string;
    'updatedAt': string;
    'oaSpec': OaSpecRef;
    'fetch': VersionFetchState;
}

export const ApiEntityManagementModeEnum = {
    Private: 'private',
    Platform: 'platform'
} as const;

export type ApiEntityManagementModeEnum = typeof ApiEntityManagementModeEnum[keyof typeof ApiEntityManagementModeEnum];